+++ src/contextHelper.ts
@@ -1,34 +1,45 @@
 import {
     Account,
+    AttributeDefinition,
+    BaseAccount,
     FormDefinitionResponseBeta,
     FormInstanceResponseBeta,
+    IdentityBeta,
     IdentityDocument,
     OwnerDto,
+    Schema,
     Source,
     WorkflowBeta,
 } from 'sailpoint-api-client'
 import { Config } from './model/config'
 import { SDKClient } from './sdk-client'
-import { AccountSchema, Context, logger } from '@sailpoint/connector-sdk'
 import {
-    buildDynamicSchema,
-    buildReviewersMap,
+    AccountSchema,
+    ConnectorError,
+    ConnectorErrorType,
+    Context,
+    SchemaAttribute,
+    logger,
+} from '@sailpoint/connector-sdk'
+import {
+    attrConcat,
+    attrSplit,
+    combineArrays,
+    composeErrorMessage,
     getExpirationDate,
     getFormName,
     getInputFromDescription,
     getOwnerFromSource,
     lm,
-    logErrors,
-    processFormInstance,
+    md,
     processUncorrelatedAccount,
-    refreshAccount,
-    sendEmail,
+    sleep,
     updateAccountLinks,
 } from './utils'
-import { WORKFLOW_NAME } from './constants'
+import { IDENTITYNOTFOUNDRETRIES, IDENTITYNOTFOUNDWAIT, WORKFLOW_NAME, reservedAttributes } from './constants'
 import { UniqueForm } from './model/form'
-import { buildUniqueAccount, buildUniqueAccountFromID, buildUniqueID } from './utils/unique'
-import { Email } from './model/email'
+import { buildUniqueAccount, buildUniqueID } from './utils/unique'
+import { Email, ErrorEmail } from './model/email'
 import { UniqueAccount } from './model/account'
 import { AxiosError } from 'axios'
 import { v4 as uuidv4 } from 'uuid'
@@ -90,18 +101,14 @@ export class ContextHelper {
         this.sources = allSources.filter((x) => this.config?.sources.includes(x.name))
 
         if (!this.source) {
-            throw new Error('No connector source was found on the tenant.')
+            throw new ConnectorError('No connector source was found on the tenant.')
         }
 
         const owner = getOwnerFromSource(this.source)
         const wfName = `${WORKFLOW_NAME} (${this.config!.cloudDisplayName})`
         this.emailer = await this.getEmailWorkflow(wfName, owner)
 
-        if (!this.emailer) {
-            throw new Error('Unable to instantiate email workflow')
-        }
-
-        this.reviewerIDs = await buildReviewersMap(this.client, this.config, this.source, this.sources)
+        this.reviewerIDs = await this.buildReviewersMap()
 
         if (!skipData) {
             this.identities = await this.listIdentities()
@@ -130,12 +137,12 @@ export class ContextHelper {
         this.errors = []
     }
 
-    private getClient(): SDKClient {
-        return this.client!
-    }
-
     getSource(): Source {
-        return this.source!
+        if (this.source) {
+            return this.source
+        } else {
+            throw new ConnectorError('No connector source was found on the tenant.')
+        }
     }
 
     listSources(): Source[] {
@@ -164,9 +171,9 @@ export class ContextHelper {
     }
 
     private async listIdentities(): Promise<IdentityDocument[]> {
-        const c = 'getIdentities'
+        const c = 'listIdentities'
         logger.info(lm('Fetching identities.', c))
-        const identities = await this.getClient().listIdentities()
+        const identities = await this.client.listIdentities()
 
         return identities ? identities : []
     }
@@ -180,12 +187,11 @@ export class ContextHelper {
     }
 
     private async listAccounts(): Promise<Account[]> {
-        const c = 'getAccounts'
-        const client = this.getClient()
+        const c = 'listAccounts'
         const source = this.getSource()
 
         logger.info(lm('Fetching existing accounts.', c))
-        let accounts = await client.listAccountsBySource(source.id!)
+        let accounts = await this.client.listAccountsBySource(source.id!)
         accounts = accounts || []
 
         logger.debug(lm('Updating existing account links.', c))
@@ -215,9 +221,7 @@ export class ContextHelper {
     }
 
     async getAccount(id: string): Promise<Account | undefined> {
-        const client = this.getClient()
-
-        const account = await client.getAccount(id)
+        const account = await this.client.getAccount(id)
 
         return account
     }
@@ -235,18 +239,16 @@ export class ContextHelper {
     }
 
     async listAuthoritativeAccounts(): Promise<Account[]> {
-        const c = 'getAuthoritativeAccounts'
-        const client = this.getClient()
-        const sources = this.listSources()
+        const c = 'listAuthoritativeAccounts'
 
         logger.info(lm('Fetching authoritative accounts.', c))
-        const authoritativeAccounts = await client.listAccounts(sources.map((x) => x.id!))
+        const authoritativeAccounts = await this.client.listAccounts(this.sources.map((x) => x.id!))
 
         return authoritativeAccounts
     }
 
     async listUniqueAccounts(): Promise<UniqueAccount[]> {
-        const c = 'getUniqueAccounts'
+        const c = 'listUniqueAccounts'
         const accounts: UniqueAccount[] = []
         const uuids: string[] = []
 
@@ -254,7 +256,7 @@ export class ContextHelper {
         for (const account of this.accounts) {
             updateAccountLinks(account, this.identities, this.config.sources)
         }
-        if (this.config?.deleteEmpty) {
+        if (this.config.deleteEmpty) {
             this.accounts = this.accounts.filter(
                 (x) =>
                     !(
@@ -289,9 +291,8 @@ export class ContextHelper {
     }
 
     async getUniqueAccount(account: Account): Promise<UniqueAccount | undefined> {
+        const c = 'getUniqueAccount'
         const sources = this.sources.map((x) => x.name)
-        const client = this.getClient()
-        const schema = await this.getSchema()
 
         let sourceAccounts: Account[] = []
         for (const sourceName of sources) {
@@ -303,14 +304,124 @@ export class ContextHelper {
         if (sourceAccounts.length === 0) sourceAccounts.push(account)
 
         try {
-            const uniqueAccount = await refreshAccount(
-                account,
-                sourceAccounts,
-                schema,
-                this.identities,
-                this.config,
-                client
-            )
+            logger.debug(lm(`Refreshing ${account.attributes!.uniqueID} account`, c, 1))
+            const schema = await this.getSchema()
+            const attributes = account.attributes
+
+            for (const attrDef of schema.attributes) {
+                if (!reservedAttributes.includes(attrDef.name)) {
+                    const attrConf = this.config.merging_map.find((x) => x.identity === attrDef.name)
+                    const attributeMerge = attrConf?.attributeMerge || this.config.attributeMerge
+                    let firstSource = true
+                    for (const sourceAccount of sourceAccounts) {
+                        let value: any
+                        if (attrConf) {
+                            for (const accountAttr of attrConf.account) {
+                                if (!sourceAccount.attributes) logger.warn(sourceAccount)
+                                value = sourceAccount.attributes![accountAttr]
+                                if (value) break
+                            }
+                        } else {
+                            value = sourceAccount.attributes![attrDef.name]
+                        }
+                        if (value) {
+                            let lst: string[]
+                            switch (attributeMerge) {
+                                case 'multi':
+                                    if (firstSource) {
+                                        lst = [].concat(value)
+                                    } else {
+                                        let previousList: string[] = [].concat(attributes![attrDef.name])
+                                        if (previousList.length === 0) {
+                                            lst = [].concat(value)
+                                        } else if (previousList.length > 1) {
+                                            lst = [...previousList, value]
+                                        } else {
+                                            lst = [...attrSplit(previousList[0]), value]
+                                        }
+                                    }
+                                    attributes![attrDef.name] = Array.from(new Set(lst))
+                                    break
+
+                                case 'concatenate':
+                                    if (firstSource) {
+                                        lst = [].concat(value)
+                                    } else {
+                                        lst = []
+                                        let previousList: string[] = [].concat(attributes![attrDef.name])
+                                        for (const item of previousList) {
+                                            lst = lst.concat(attrSplit(item))
+                                        }
+                                        lst = lst.concat(attrSplit(value))
+                                    }
+                                    attributes![attrDef.name] = attrConcat(lst)
+                                    break
+                                case 'first':
+                                    if (firstSource) {
+                                        attributes![attrDef.name] = value
+                                    }
+                                    break
+
+                                case 'source':
+                                    const source = attrConf?.source
+                                    if (sourceAccount.sourceName === source) {
+                                        attributes![attrDef.name] = value
+                                    }
+                                    break
+                                default:
+                                    break
+                            }
+                        }
+                        firstSource = false
+                    }
+                }
+            }
+
+            attributes!.status = Array.from(new Set(attributes!.status))
+
+            if (account.uncorrelated) {
+                logger.debug(lm(`New account. Needs to be enabled.`, c, 2))
+            } else {
+                logger.debug(lm(`Existing account. Enforcing defined correlation.`, c, 1))
+                let identity: IdentityDocument | IdentityBeta | undefined
+                let accounts: Account[] | BaseAccount[]
+                identity = this.identities.find((x) => x.id === account.identityId) as IdentityDocument
+                if (!identity) {
+                    let count = 0
+                    let wait = IDENTITYNOTFOUNDWAIT
+                    while (!identity) {
+                        identity = await this.client.getIdentity(account.identityId!)
+                        if (!identity) {
+                            if (++count > IDENTITYNOTFOUNDRETRIES)
+                                throw new Error(
+                                    `Identity ${account.identityId} for account ${account.nativeIdentity} not found`
+                                )
+
+                            logger.warn(lm(`Identity ID ${account.identityId} not found. Re-trying...`, c, 1))
+                            await sleep(wait)
+                            wait = wait + IDENTITYNOTFOUNDWAIT
+                        }
+                    }
+                    accounts = await this.client.getAccountsByIdentity(identity!.id!)
+                } else {
+                    accounts = (identity as IdentityDocument).accounts!
+                }
+
+                for (const acc of account.attributes!.accounts as string[]) {
+                    const uid: string = (identity.attributes as any).uid
+                    try {
+                        if (!accounts.find((x) => x.id === acc)) {
+                            logger.debug(lm(`Correlating ${acc} account with ${uid}.`, c, 1))
+                            const response = await this.client.correlateAccount(identity.id as string, acc)
+                        }
+                    } catch (e) {
+                        logger.error(lm(`Failed to correlate ${acc} account with ${uid}.`, c, 1))
+                        account.attributes!.accounts = account.attributes!.accounts.filter((x: string) => x !== acc)
+                    }
+                }
+            }
+
+            const uniqueAccount = new UniqueAccount(account, schema)
 
             return uniqueAccount
         } catch (error) {
@@ -348,36 +459,28 @@ export class ContextHelper {
     }
 
     async listForms(): Promise<FormDefinitionResponseBeta[]> {
-        const client = this.getClient()
-
-        const forms = await client.listForms()
+        const forms = await this.client.listForms()
         const currentForms = forms.filter((x) => x.name?.startsWith(this.getFormName()))
 
         return currentForms
     }
 
     async createUniqueForm(form: UniqueForm): Promise<FormDefinitionResponseBeta> {
-        const client = this.getClient()
-
-        const response = await client.createForm(form)
+        const response = await this.client.createForm(form)
         this.forms.push(response)
 
         return response
     }
 
     async deleteForm(form: FormDefinitionResponseBeta) {
-        const client = this.getClient()
-
-        await client.deleteForm(form.id!)
+        await this.client.deleteForm(form.id!)
 
         const index = this.forms.findIndex((x) => x.id === form.id!)
         this.forms.splice(index, 1)
     }
 
     async getFormInstances(forms?: FormDefinitionResponseBeta[]): Promise<FormInstanceResponseBeta[]> {
-        const client = this.getClient()
-
-        let formInstances = await client.listFormInstances()
+        let formInstances = await this.client.listFormInstances()
         //Order from older to newer
         formInstances = formInstances.sort((a, b) => new Date(a.modified!).valueOf() - new Date(b.modified!).valueOf())
 
@@ -394,24 +497,21 @@ export class ContextHelper {
         this.formInstances.splice(this.formInstances.indexOf(formInstance), 1)
     }
 
-    async processFormInstance(
-        formInstance: FormInstanceResponseBeta
-    ): Promise<{ decision: string; account: string; message: string }> {
-        return processFormInstance(this.getClient(), formInstance)
-    }
+    // async processFormInstance(
+    //     formInstance: FormInstanceResponseBeta
+    // ): Promise<{ decision: string; account: string; message: string }> {
+    //     return processFormInstance(this.client, formInstance)
+    // }
 
     async createFormInstance(form: FormDefinitionResponseBeta, reviewerID: string) {
-        const client = this.getClient()
-        const source = this.getSource()
-
         const expire = getExpirationDate(this.config)
         const formInput = form.formInput?.reduce(getInputFromDescription, {})
 
-        const currentFormInstance = await client.createFormInstance(
+        const currentFormInstance = await this.client.createFormInstance(
             form.id!,
             formInput!,
             [reviewerID],
-            source.id!,
+            this.source!.id!,
             expire
         )
         this.formInstances.push(currentFormInstance)
@@ -426,14 +526,13 @@ export class ContextHelper {
     async processUncorrelatedAccount(
         uncorrelatedAccount: Account
     ): Promise<{ processedAccount: Account | undefined; uniqueForm: UniqueForm | undefined }> {
-        const source = this.getSource()
         const deduplicate = await this.isMergingEnabled()
 
         const response = await processUncorrelatedAccount(
             uncorrelatedAccount,
             this.accounts,
             this.currentIdentities,
-            source,
+            this.source!,
             this.config,
             deduplicate
         )
@@ -450,20 +549,40 @@ export class ContextHelper {
     }
 
     async buildUniqueAccountFromID(id: string): Promise<UniqueAccount> {
-        const client = this.getClient()
-        const source = this.getSource()
         const schema = await this.getSchema()
 
-        const uniqueAccount = await buildUniqueAccountFromID(id, schema, source, this.identities, this.config, client)
+        const c = 'buildUniqueAccountFromID'
+        logger.debug(lm(`Fetching original account`, c, 1))
+        const account = await this.client.getAccountBySourceAndNativeIdentity(this.getSource().id!, id)
+        const sourceAccounts: Account[] = []
+        if (account) {
+            const identity = await this.client.getIdentity(account.identityId!)
+            const accounts = await this.client.getAccountsByIdentity(identity!.id!)
+            const correlatedAccounts = accounts
+                .filter((x) => this.config.sources.includes(x.sourceName!))
+                .map((x) => x.id as string)
+            account.attributes!.accounts = combineArrays(correlatedAccounts, account.attributes!.accounts)
 
-        return uniqueAccount
+            for (const acc of account.attributes!.accounts) {
+                logger.debug(lm(`Looking for ${acc} account`, c, 1))
+                const response = await this.client.getAccount(acc)
+                if (response) {
+                    logger.debug(lm(`Found linked account ${response.name} (${response.sourceName})`, c, 1))
+                    sourceAccounts.push(response)
+                } else {
+                    logger.error(lm(`Unable to find account ID ${acc}`, c, 1))
+                }
+            }
+
+            const uniqueAccount = await this.getUniqueAccount(account)
+            return uniqueAccount!
+        } else {
+            throw new ConnectorError('Account not found', ConnectorErrorType.NotFound)
+        }
     }
 
     async buildUniqueID(id: string): Promise<string> {
-        const client = this.getClient()
-        const source = this.getSource()
-
-        const account = await client.getAccountBySourceAndNativeIdentity(source.id!, id)
+        const account = await this.client.getAccountBySourceAndNativeIdentity(this.source!.id!, id)
         const uniqueID = await buildUniqueID(account!, this.ids, this.config)
         this.ids.push(uniqueID)
 
@@ -471,9 +590,7 @@ export class ContextHelper {
     }
 
     async sendEmail(email: Email) {
-        const client = this.getClient()
-
-        await sendEmail(email, this.emailer!, client)
+        await this.client.testWorkflow(this.emailer!.id!, email)
     }
 
     loadSchema(schema: AccountSchema) {
@@ -481,14 +598,11 @@ export class ContextHelper {
     }
 
     async getSchema(): Promise<AccountSchema> {
-        const client = this.getClient()
-        const sources = this.listSources()
-
         let schema: AccountSchema
         if (this.schema) {
             schema = this.schema
         } else {
-            schema = await buildDynamicSchema(sources, this.config, client)
+            schema = await this.buildDynamicSchema()
         }
 
         return schema
@@ -496,16 +610,6 @@ export class ContextHelper {
 
     getEmailer(): WorkflowBeta {
         return this.emailer!
-    }
-
-    async logErrors(context: Context, input: any) {
-        const source = this.getSource()
-        const workflow = this.getEmailer()
-        const client = this.getClient()
-
-        if (this.errors.length > 0) {
-            await logErrors(context, input, this.errors, source, workflow, client)
-        }
     }
 
     async fetchUniqueIDs() {
@@ -531,7 +635,230 @@ export class ContextHelper {
             workflow = await this.client.createWorkflow(emailWorkflow)
         }
 
+        if (!workflow) throw new Error('Unable to instantiate email workflow')
+
         return workflow
+    }
+
+    private async buildReviewersMap(): Promise<Map<string, string[]>> {
+        const reviewersMap = new Map<string, string[]>()
+        let defaultReviewerIDs: string[] = []
+        if (!this.config.merging_reviewerIsSourceOwner) {
+            defaultReviewerIDs = await this.fetchReviewerIDs(this.source!)
+        }
+
+        for (const source of this.sources) {
+            if (this.config.merging_reviewerIsSourceOwner) {
+                const reviewerIDs = await this.fetchReviewerIDs(source)
+                reviewersMap.set(source.name, reviewerIDs)
+            } else {
+                reviewersMap.set(source.name, defaultReviewerIDs)
+            }
+        }
+
+        return reviewersMap
+    }
+
+    private async fetchReviewerIDs(source: Source): Promise<string[]> {
+        const c = 'fetchReviewerIDs'
+        logger.debug(lm(`Fetching reviewers for ${source.name}`, c, 1))
+        let reviewers: string[] = []
+
+        if (source.managementWorkgroup) {
+            logger.debug(lm(`Reviewer is ${source.managementWorkgroup.name} workgroup`, c, 1))
+            const workgroups = await this.client.listWorkgroups()
+            const workgroup = workgroups.find((x) => x.id === source.managementWorkgroup!.id)
+            if (workgroup) {
+                logger.debug(lm('Workgroup found', c, 1))
+                const members = await this.client.listWorkgroupMembers(workgroup.id!)
+                reviewers = members.map((x) => x.id!)
+            }
+        } else if (source.owner || reviewers.length === 0) {
+            logger.debug(lm('Reviewer is the owner', c, 1))
+            const reviewerIdentity = await this.client.getIdentity(source.owner.id!)
+            if (reviewerIdentity) {
+                logger.debug(lm('Reviewer found', c, 1))
+                reviewers.push(reviewerIdentity.id!)
+            } else {
+                logger.error(lm(`Reviewer not found ${source.owner.name}`, c, 1))
+            }
+        } else {
+            logger.warn(lm(`No reviewer provided. Merging forms will not be processed.`, c, 1))
+        }
+
+        return reviewers
+    }
+
+    async processFormInstance(
+        formInstance: FormInstanceResponseBeta
+    ): Promise<{ decision: string; account: string; message: string }> {
+        const c = 'processFormInstance'
+        const now = new Date().toISOString()
+        let message = ''
+        const decision = formInstance.formData!['identities'].toString()
+        const account = (formInstance.formInput!['account'] as any).value
+        const reviewerIdentity = await this.client.getIdentityBySearch(formInstance.recipients![0].id!)
+        const reviewerName = reviewerIdentity
+            ? reviewerIdentity.displayName
+                ? reviewerIdentity.displayName
+                : reviewerIdentity.name
+            : formInstance.recipients![0].id!
+
+        if (decision === 'This is a new identity') {
+            message = `New identity approved by ${reviewerName}`
+        } else {
+            const source = (formInstance.formInput!.source as any).value
+            message = `Assignment approved by ${reviewerName}`
+        }
+
+        return { decision, account, message }
+    }
+
+    private async buildDynamicSchema(): Promise<AccountSchema> {
+        const c = 'buildDynamicSchema'
+        logger.debug(lm('Fetching sources.', c, 1))
+        const schemas: Schema[] = []
+        logger.debug(lm('Fetching schemas.', c, 1))
+        for (const source of this.sources) {
+            const sourceSchemas = await this.client.listSourceSchemas(source.id!)
+            schemas.push(sourceSchemas.find((x) => x.name === 'account') as Schema)
+        }
+
+        logger.debug(lm('Compiling attributes.', c, 1))
+        let combinedAttributes: Map<string, AttributeDefinition> = new Map()
+        for (const schema of schemas.reverse()) {
+            schema.attributes?.forEach((x) => combinedAttributes.set(x.name!, x))
+        }
+
+        logger.debug(lm('Defining static attributes.', c, 1))
+        const attributes: SchemaAttribute[] = [
+            {
+                name: 'uniqueID',
+                description: 'Unique ID',
+                type: 'string',
+                required: true,
+            },
+            {
+                name: 'uuid',
+                description: 'UUID',
+                type: 'string',
+                required: true,
+            },
+            {
+                name: 'history',
+                description: 'History',
+                type: 'string',
+                multi: true,
+            },
+            {
+                name: 'status',
+                description: 'Status',
+                type: 'string',
+                multi: true,
+                entitlement: true,
+                managed: false,
+                schemaObjectType: 'status',
+            },
+            {
+                name: 'accounts',
+                description: 'Account IDs',
+                type: 'string',
+                multi: true,
+                entitlement: false,
+            },
+            {
+                name: 'reviews',
+                description: 'Reviews',
+                type: 'string',
+                multi: true,
+                entitlement: false,
+            },
+        ]
+
+        logger.debug(lm('Processing attribute merge mapping.', c, 1))
+        for (const mergingConf of this.config.merging_map) {
+            const description = mergingConf.source ? mergingConf.source : mergingConf.identity
+            const attribute: any = {
+                name: mergingConf.identity,
+                description,
+                type: 'string',
+            }
+
+            switch (mergingConf.attributeMerge) {
+                case 'multi':
+                    attribute.multi = true
+                    attribute.entitlement = true
+                    break
+
+                case 'concatenate':
+                    attribute.multi = false
+                    break
+
+                default:
+                    break
+            }
+
+            attributes.push(attribute)
+        }
+
+        logger.debug(lm('Processing existing attributes.', c, 1))
+        for (const attribute of combinedAttributes.values()) {
+            if (!attributes.find((x) => x.name === attribute.name!)) {
+                const mergingConf = this.config.merging_map.find((x) => x.attributeMerge?.includes(attribute.name!))
+                let attributeMerge: string
+                if (mergingConf?.attributeMerge) {
+                    attributeMerge = mergingConf.attributeMerge
+                } else {
+                    attributeMerge = this.config.attributeMerge
+                }
+                const matchingSchemas = schemas.filter((x) => x.attributes?.find((y) => y.name === attribute.name))
+                switch (attributeMerge) {
+                    case 'multi':
+                        if (matchingSchemas.length > 1) {
+                            attribute.isMulti = true
+                            attribute.type = 'STRING'
+                        }
+                        break
+
+                    case 'concatenate':
+                        attribute.isMulti = false
+                        attribute.type = 'STRING'
+                        break
+
+                    default:
+                        break
+                }
+
+                if (attribute.isMulti) {
+                    attribute.isEntitlement = true
+                    attribute.isGroup = false
+                }
+
+                const description = (
+                    attribute.description === null || attribute.description === ''
+                        ? attribute.name
+                        : attribute.description
+                ) as string
+                const schemaAttribute: SchemaAttribute = {
+                    name: attribute.name!,
+                    description,
+                    type: attribute.type ? attribute.type.toLowerCase() : 'string',
+                    multi: attribute.isMulti,
+                    managed: false,
+                    entitlement: attribute.isEntitlement,
+                }
+
+                attributes.push(schemaAttribute)
+            }
+        }
+
+        const schema: any = {
+            attributes,
+            displayAttribute: 'uuid',
+            identityAttribute: 'uuid',
+        }
+
+        return schema
     }
 
     handleError(error: any) {
@@ -548,5 +875,17 @@ export class ContextHelper {
         logger.error(message)
         logger.error(error)
         this.errors.push(message)
+    }
+
+    async logErrors(context: Context, input: any) {
+        if (this.errors.length > 0) {
+            const message = composeErrorMessage(context, input, this.errors)
+
+            const ownerID = this.getSource().owner.id as string
+            const recipient = await this.client.getIdentityBySearch(ownerID)
+            const email = new ErrorEmail(this.getSource(), recipient!.email!, message)
+
+            await this.sendEmail(email)
+        }
     }
 }
